Completed initial design
This design uses typical values of a APM system gathered from the literature to develop a prototype. Physics is applied to show the magnetic field. It ignores thermal effects on the system. Furthermore the magnetic arrangement has alot of room for improvement. It requires directional tuning and shape modification to maximize the surface area. This will have a direct impact on the decceleration.